pyproject.toml
[project]
name = "coldfront-exporter-v1"
version = "0.0.1"
description = "Export ColdFront database to YAML files"
readme = "README.md"
authors = [
    { name = "Andrew E. Bruno", email = "aebruno2@buffalo.edu" }
]
license = "Apache-2.0"
license-files = ["LICENSES/*.txt"]
keywords = ["high-performance-computing", "resource-allocation"]
classifiers = [
    "Framework :: Django",
    "Environment :: Plugins",
    'Topic :: Scientific/Engineering',
    'Topic :: System :: Systems Administration',
]
requires-python = ">=3.12"
dependencies = [
    "coldfront>=1.1.7,<2",
    "djangorestframework-yaml>=2.0.0",
]

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "reuse>=6.2.0",
    "ruff>=0.15.7",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
